Redesign the selector definition on reports.
This is a bit simpler, and more closely represents the data in use by
the report edit form (which is not yet updated to work with these
changes).

We should be able to dispense with a lot of the definition<->form
conversion code we needed before. That's the next commit...