Add Check-in admin feature.
This has been brought over from an old plugin we've been using at Cedar
Ridge for close to 5 years. It's a bit rough around the edges, but I
chose to include it anyway and iterate.

For this to be useful, the API must be used to check people in. In the
near future I hope to release the desktop touch-screen application that
utilizes this API and allows for self-check-in kiosks.