Update image? method
We have run into a situation where this method generates an exception
when being run against a particular PDF that one of the members has
uploaded.  The actual issue lies w/ ImageMagick's identify, which
minimagick calls.  Note that we experience this issue in production
using ImageMagick 6.7.7-10, but not 6.8.9-7 which we use in development.

This change simply moves some of the functionaly in the image? method
to a new method and wraps it with some exception handling to handle
this particular failure scenario.

Closes #283